+++ types/components/charts/BarChart.d.ts
@@ -5,6 +5,11 @@ import BarChartOptions from '~/interfaces/BarChartOptions';
 export default class {
     #private;
     /**
+     * Read-only ApexCharts instance.
+     * @type {ApexCharts|undefined}
+     */
+    get instance(): ApexCharts | undefined;
+    /**
      * Create a new instance of the BarChart class.
      * @param {string|HTMLElement|JQuery} element HTMLElement selector, element, or JQuery object.
      * @param {BarChartOptions} options Chart options.